refactor: update prose utility
